refactor: move conversation dto to conversation package
